LayoutTests/platform/chromium/fast/forms/calendar-picker/calendar-picker-pre-100-year.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../../../fast/js/resources/js-test-pre.js"></script>
<script src="../../../../../fast/forms/resources/common.js"></script>
<script src="../../../../../fast/forms/resources/picker-common.js"></script>
<script src="resources/calendar-picker-common.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<input type=date id=date value="0002-02-02">
<script>
description('Check if year before 100 can be set.');

openPicker(document.getElementById('date'), finishTest);

function finishTest() {
    var daysContainer = popupWindow.document.getElementsByClassName("days-area")[0];
    sendKey(daysContainer, 'Right');
    sendKey(daysContainer, 'Enter');
    shouldBe('document.getElementById("date").value', '"0002-02-03"');
    finishJSTest();
}
</script>
<script src="../../../../../fast/js/resources/js-test-post.js"></script>
</body>
</html>
